moved away from deprecated plugin-jsx
