+++ .gitignore
@@ -1,1 +1,5 @@
 .idea
+
+.DS_Store
+
+jspm_packages
