+++ src/app/index.js
@@ -1,8 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import {Provider} from 'react-redux'
-import {store, socket} from 'app/store.js'
-import Form from 'app/form.jsx!'
+import {store, socket} from './store.js'
+import Form from './form.js'
 
 //TODO: Dispatch whatever the socket emits
