+++ src/index.html
@@ -10,7 +10,7 @@
 <body>
 <div id="root">Loading...</div>
 <script src="jspm_packages/system.js"></script>
-<script src="config.js"></script>
+<script src="jspm.config.js"></script>
 <script>
     SystemJS.load('app/index.js')
 </script>
